+++ test/blind_index_test.rb
@@ -49,6 +49,13 @@ class BlindIndexTest < Minitest::Test
     assert User.where(email: "test@example.org").first
   end
 
+  def test_where_table
+    skip if mongoid? || ActiveRecord::VERSION::STRING.to_f < 5.2
+
+    create_user
+    assert User.where(users: {email: "test@example.org"}).first
+  end
+
   def test_where_array
     create_user
     create_user(email: "test2@example.org")